update 12/10 21:36

+++ routes/comment.js
@@ -1,11 +1,13 @@
-const express = require("express");
-const Comment = require("../models/Comments");
-//const Like = require("../models/likes");
-const middleware = require("../middleware/auth-middleware");
+const express = require('express');
+const Comment = require('../models/Comments');
+const User = require('../models/User');
+const middleware = require('../middleware/auth-middleware');
 const router = express.Router();
 
+
+
 // 댓글 작성
-router.post('/comment',  middleware, async (req, res) => {
+router.post('/',  middleware, async (req, res) => {
   try{
     //const post= req.body
     const { postId } = req.body
@@ -21,9 +23,10 @@ router.post('/comment',  middleware, async (req, res) => {
     if (recentComment.length != 0) {
       commentId = recentComment[0]["commentId"] + 1
     }
-    
+    console.log(userId)
+    console.log({userId})
     await Comment.create({ commentId, postId, userId, commentVal })
-    res.send({ result: "success "})
+    res.send({ result: "success"})
   }catch(error){
     res.send(error)
     console.log(error);
@@ -32,33 +35,48 @@ router.post('/comment',  middleware, async (req, res) => {
 })
 
 // 댓글 불러오기
-router.get("/comment/:commentId", async (req, res) => {
-  const { postId } = req.params
+router.post("/detail", async (req, res) => {
+  const { postId } = req.body
   const comment = await Comment.find({ postId }).sort("-commentId")
   res.json({ comment: comment })
 })
-
 // 댓글 삭제
-router.delete('/comment',middleware, async (req, res)=>{
-  const { commentId } = req.body;
-  try{
-      await Comment.deleteOne({ commentId });
-      res.send({succes:true})
-  }catch(error){
-      res.status(400).send({errormessage:"삭제중 오류가 발생했습니다."})
-      console.log(error);
+router.delete('/detail', middleware, async (req, res)=>{
+  const { commentId } = req.body
+  const { userId } = req.body
+
+  const current_user = userId
+  // console.log("current_user", current_user)
+  const userComment = await Comment.findOne({ commentId })
+  // console.log("userComment",userComment)
+  const commentUserId = userComment["userId"]
+  // console.log("commentUserId", commentUserId)
+
+  if (current_user === commentUserId) {
+    await Comment.deleteOne({ commentId })
+    res.send({succes: true})
+  } else {
+    res.send({result: '로그인 후 이용하시오'})
   }
 })
 
 // 댓글 수정
-router.put('/comment',middleware, async (req, res)=>{
-  const { commentId, commentVal } = req.body;
-  try{
-      await Comment.updateOne({ commentId }, { $set: { commentVal }})
-      res.send({succes:true})
-  }catch(error){
-      res.status(400).send({errormessage:"수정중 오류가 발생했습니다."})
-      console.log(error);
+router.put('/detail',middleware, async (req, res)=>{
+  const { commentId, commentVal } = req.body
+  const { userId } = req.body
+
+  const current_user = userId
+  // console.log("current_user", current_user)
+  const userComment = await Comment.findOne({ commentId })
+  // console.log("userComment",userComment)
+  const commentUserId = userComment["userId"]
+  // console.log("commentUserId", commentUserId)
+
+  if (current_user === commentUserId) {
+    await Comment.updateOne({ commentId }, { $set: { commentVal }})
+    res.send({succes: true})
+  } else {
+    res.send({errormessage:"수정중 오류가 발생했습니다."})
   }
 })
